+++ common/src/main/java/com/gregtechceu/gtceu/common/data/GTPlacements.java
@@ -5,14 +5,12 @@ import com.gregtechceu.gtceu.api.data.worldgen.BiomeWeightModifier;
 import com.gregtechceu.gtceu.api.data.worldgen.generator.BiomePlacement;
 import com.gregtechceu.gtceu.api.data.worldgen.generator.VeinCountFilter;
 import com.gregtechceu.gtceu.data.recipe.CustomTags;
-import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderGetter;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BootstapContext;
 import net.minecraft.data.worldgen.placement.PlacementUtils;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.level.biome.Biome;
-import net.minecraft.world.level.levelgen.blockpredicates.BlockPredicate;
 import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
 import net.minecraft.world.level.levelgen.placement.*;
 
@@ -32,16 +30,16 @@ public class GTPlacements {
         HolderGetter<Biome> biomeLookup = ctx.lookup(Registries.BIOME);
 
         PlacementUtils.register(ctx, RUBBER_CHECKED, featureLookup.getOrThrow(GTConfiguredFeatures.RUBBER),
-                PlacementUtils.filteredByBlockSurvival(GTBlocks.RUBBER_SAPLING.get()), new BiomePlacement(List.of(
+                new BiomePlacement(List.of(
                         new BiomeWeightModifier(() -> biomeLookup.getOrThrow(CustomTags.IS_SWAMP), 50)
                 )),
-                PlacementUtils.countExtra(0, 0.005F, 1),
+                PlacementUtils.countExtra(0, 0.05F, 1),
                 InSquarePlacement.spread(),
                 SurfaceWaterDepthFilter.forMaxDepth(0),
                 PlacementUtils.HEIGHTMAP_OCEAN_FLOOR,
-                BlockPredicateFilter.forPredicate(BlockPredicate.wouldSurvive(GTBlocks.RUBBER_SAPLING.getDefaultState(), BlockPos.ZERO)),
-                BiomeFilter.biome()
-        );
+                BiomeFilter.biome(),
+                PlacementUtils.filteredByBlockSurvival(GTBlocks.RUBBER_SAPLING.get())
+                );
         PlacementUtils.register(ctx, ORE, featureLookup.getOrThrow(GTConfiguredFeatures.ORE), VeinCountFilter.count(), InSquarePlacement.spread());
     }
 }