+++ common/src/main/java/com/gregtechceu/gtceu/data/tags/IBiomeTagsProvider.java
@@ -1,17 +1,18 @@
 package com.gregtechceu.gtceu.data.tags;
 
+import com.gregtechceu.gtceu.data.recipe.CustomTags;
 import net.minecraft.data.tags.TagsProvider;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.level.biome.Biome;
+import net.minecraft.world.level.biome.Biomes;
 
 public interface IBiomeTagsProvider<T extends TagsProvider.TagAppender<Biome>> {
 
     default void generateTags() {
-        //TODO what the hell?
-//        tag(CustomTags.IS_SWAMP, Biomes.SWAMP, Biomes.MANGROVE_SWAMP);
-//        addRubberTreeTag();
-//        addSandyTag();
+        tag(CustomTags.IS_SWAMP).addOptional(Biomes.SWAMP.location()).addOptional(Biomes.MANGROVE_SWAMP.location());
+        addRubberTreeTag();
+        addSandyTag();
     }
 
     void addRubberTreeTag();